+++ Trader.h
@@ -13,7 +13,8 @@ private:
 public: 
 	Trader(int id, String traderName, char traderType);
 	Bid generateBid();
-	Bid generateBid(double MINPRICE, double MAXPRICE, int MINQUANTITY, int MAXQUANTITY);
+	Bid generateBid(int MINPRICE, int MAXPRICE, int MINQUANTITY, int MAXQUANTITY);
 	int getId();
+	
 };
